First draft of Week 3 lecture
For any remaining time after each of these lectures, we will train students in the basics of SQL. As long as the SQL syntax necessary to complete the following week's assignment has been covered, the lecturer can introduce the students to as much of the language as they feel comfortable with. I'm thinking of just having one giant "SQL Basics" lecture that the instructor can go back to each week and pick up where they left off. Essentially, the critical thinking / pitfalls lectures and the SQL lectures can be taught in parallel.