First draft of the Recidivism lecture
This one needs help. I'm not really happy with the generalities in the final slides (pretty much copied from the ADSA website). Let's hit up our data science ethics resources to find more specific guidelines that would help with this specific case. Rather than provide some general ethical framework, I'd like to come up with a concrete list of steps that a student well-versed in ethical frameworks could follow. Once we're satisfied that the advice is helpful, we could say more about the framework, because it's "proved itself" in a sense.